+++ packages/gestalt/src/SegmentedControl.test.js
@@ -1,7 +1,6 @@
 // @flow
 import React from 'react';
 import { create } from 'react-test-renderer';
-import { shallow } from 'enzyme';
 import SegmentedControl from './SegmentedControl.js';
 
 test('SegmentedControl renders', () => {
@@ -25,24 +24,4 @@ test('SegmentedControl with responsive widths renders', () => {
     />
   ).toJSON();
   expect(tree).toMatchSnapshot();
-});
-
-test('SegmentedControl handles click', () => {
-  const mockOnChange = jest.fn();
-  const wrapper = shallow(
-    <SegmentedControl
-      items={['Item']}
-      selectedItemIndex={0}
-      onChange={mockOnChange}
-    />
-  );
-
-  const mockEvent = {};
-  wrapper.find('button').simulate('click', mockEvent);
-
-  expect(mockOnChange).toHaveBeenCalled();
-  expect(mockOnChange).toHaveBeenCalledWith({
-    event: mockEvent,
-    activeIndex: 0,
-  });
 });