+++ packages/gestalt/src/TextField.test.js
@@ -1,21 +1,19 @@
 // @flow
 import React from 'react';
 import { create } from 'react-test-renderer';
-import { shallow } from 'enzyme';
-import FormErrorMessage from './FormErrorMessage.js';
 import TextField from './TextField.js';
 
 describe('TextField', () => {
   it('Renders an FormErrorMessage if an error message is passed in', () => {
-    const wrapper = shallow(
-      <TextField errorMessage="test" id="test" onChange={jest.fn()} />
+    const component = create(
+      <TextField errorMessage="Error message" id="test" onChange={jest.fn()} />
     );
-    expect(wrapper.find(FormErrorMessage)).toHaveLength(1);
+    expect(JSON.stringify(component.toJSON())).toContain('Error message');
   });
 
   it('Does not render an FormErrorMessage when errorMessage is null', () => {
-    const wrapper = shallow(<TextField id="test" onChange={jest.fn()} />);
-    expect(wrapper.find(FormErrorMessage)).toHaveLength(0);
+    const component = create(<TextField id="test" onChange={jest.fn()} />);
+    expect(JSON.stringify(component.toJSON())).not.toContain('Error message');
   });
 
   it('TextField normal', () => {
@@ -31,7 +29,7 @@ describe('TextField', () => {
   });
 
   it('TextField with error', () => {
-    const tree = shallow(
+    const tree = create(
       <TextField
         errorMessage="error message"
         id="test"
@@ -39,12 +37,12 @@ describe('TextField', () => {
         onFocus={jest.fn()}
         onBlur={jest.fn()}
       />
-    ).html();
+    ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('TextField with hasError', () => {
-    const wrapper = shallow(
+    const tree = create(
       <TextField
         hasError
         id="test"
@@ -52,116 +50,33 @@ describe('TextField', () => {
         onFocus={jest.fn()}
         onBlur={jest.fn()}
       />
-    );
-    expect(wrapper.html()).toMatchSnapshot();
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
   });
 
   it('TextField with name', () => {
-    const tree = shallow(
+    const tree = create(
       <TextField
         name="email"
-        id="email"
+        id="test"
         onChange={jest.fn()}
         onFocus={jest.fn()}
         onBlur={jest.fn()}
       />
-    ).html();
+    ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('TextField with autocomplete', () => {
-    const tree = shallow(
+    const tree = create(
       <TextField
-        autoComplete="on"
-        id="email"
-        onChange={jest.fn()}
-        onFocus={jest.fn()}
-        onBlur={jest.fn()}
-      />
-    ).html();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('TextField with errorMessage prop change', () => {
-    const tree = shallow(
-      <TextField
+        name="email"
         id="test"
         onChange={jest.fn()}
         onFocus={jest.fn()}
         onBlur={jest.fn()}
       />
-    );
-    expect(tree.find(FormErrorMessage)).toHaveLength(0);
-    tree.setProps({
-      errorMessage: 'error message',
-    });
-    expect(tree.find(FormErrorMessage)).toHaveLength(1);
-  });
-
-  it('TextField with type number', () => {
-    const tree = shallow(
-      <TextField
-        id="test"
-        onChange={jest.fn()}
-        onFocus={jest.fn()}
-        onBlur={jest.fn()}
-        type="number"
-      />
-    ).html();
+    ).toJSON();
     expect(tree).toMatchSnapshot();
-  });
-
-  it('handles blur events', () => {
-    const mockBlur = jest.fn();
-    const tree = shallow(
-      <TextField id="test" onBlur={mockBlur} onChange={jest.fn()} />
-    );
-    tree
-      .find('input')
-      .simulate('blur', { currentTarget: { value: 'fake value' } });
-    expect(mockBlur).toHaveBeenCalledWith({
-      event: { currentTarget: { value: 'fake value' } },
-      value: 'fake value',
-    });
-  });
-
-  it('handles change events', () => {
-    const mockChange = jest.fn();
-    const tree = shallow(<TextField id="test" onChange={mockChange} />);
-    tree
-      .find('input')
-      .simulate('change', { currentTarget: { value: 'fake value' } });
-    expect(mockChange).toHaveBeenCalledWith({
-      event: { currentTarget: { value: 'fake value' } },
-      value: 'fake value',
-    });
-  });
-
-  it('handles focus events', () => {
-    const mockFocus = jest.fn();
-    const tree = shallow(
-      <TextField id="test" onChange={jest.fn()} onFocus={mockFocus} />
-    );
-    tree
-      .find('input')
-      .simulate('focus', { currentTarget: { value: 'fake value' } });
-    expect(mockFocus).toHaveBeenCalledWith({
-      event: { currentTarget: { value: 'fake value' } },
-      value: 'fake value',
-    });
-  });
-
-  it('handles key down events', () => {
-    const mockKeyDown = jest.fn();
-    const tree = shallow(
-      <TextField id="test" onChange={jest.fn()} onKeyDown={mockKeyDown} />
-    );
-    tree
-      .find('input')
-      .simulate('keyDown', { currentTarget: { value: 'fake value' } });
-    expect(mockKeyDown).toHaveBeenCalledWith({
-      event: { currentTarget: { value: 'fake value' } },
-      value: 'fake value',
-    });
   });
 });