+++ CHANGELOG.md
@@ -7,10 +7,12 @@
   </summary>
 
 ### Minor
-- IconButton: Allow `orange` icon color prop to be passed as a value (#602)
+
 - Pog: Add `orange` icon color prop to be passed as a value (#602)
 - Icon: Add directional-arrow-left icon (#605)
 - Icon: Add directional-arrow-right icon (#605)
+- IconButton: Allow `orange` icon color prop to be passed as a value (#602)
+- Video: Improve video seeking (#601)
 
 ### Patch
 
@@ -20,7 +22,7 @@
 - Internal: Add `react-testing-library` (#598)
 - Internal: Enable `react-hooks/exhaustive-deps` lint rule (#598)
 - Internal: Remove `concurrently` dependency (#599)
-- Video: Improve video seeking (#601)
+- Internal: Convert tests to React Testing Library (#603)
 
 </details>